config/requirements/test.txt
coverage
django-debug-toolbar==2.2

